(#9595) Autoflush log file destinations by default
Previously, puppet did not call `IO#flush` after writing each log
message to its log file destination. As a result, if the agent
terminated unexpectedly, any data buffered in ruby, but not yet
flushed would be lost.

This commit changes the default value of autoflush to be true, which
will cause ruby to flush its buffer. Since this property is only used
by the log file destination, and the documentation for the property
says, 'Whether log files should always flush to disk.', it seems safe
to change the default value.