Fix Rails compatibility code to be less stupid.
When the Inventory ActiveRecord node object was implemented, an alias was
created to make the `scope` vs `named_scope` change over the Rails 2.x to 3.x
releases work nicely.

Sadly, this did this by using the older name, and creating an alias on the
modern releases.  That works great until the 3.2.x release, where the
internals changed and the compatibility stuff broke.

Instead, we apply the change so that the modern version works for older
releases, which should be much more robust - not least because they don't
develop on the older versions, so they are not a moving target.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>