Clean up and unify scratch database setup logic.
Now that we have a pile of places all trying to do the same setup and
exclusion for SQLite in their tests, it pays to push all that into a central
location and make it uniformly available to all the tests.

This provides a nice, global method that can be invoked by any test suite to
configure the database, and then uses that everywhere that was doing scratch
database setup.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>