Make PuppetSpec::Files methods available at module level.
Previously you could only access the `tmpfile` and `tmpdir` helpers by mixing
the module in somewhere, despite them exclusively mutating global state.

This became inconvenient when trying to centralize the database setup, so I
moved them to module functions and made the helpers that mix in defer to the
module level methods.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>