Disable replace_file on Windows
Our current approach of setting owner/mode in the Puppet way risks
losing any extra permissions beyond owner/group/everyone, which is
significant enough that it is better to simply not support this
operation until we have a better solution. replace_file isn't used in
any code which actually runs on Windows at the moment, so this is
a benign change.