(#1168) Use implicit lower case value of certname
Without this patch we explicitly fail hard when the certname setting is
configured to something that is mixed case or upper case.

This is a problem on Windows where the ComputerName defaults to an upper
case NETBIOS name.  We cannot work around the problem in the MSI
installer effectively because we have virtually no string manipulation
capabilities and embedding JavaScript into an MSI is ill advised for
long term supportability and compatibility with antivirus software.

This patch fixes the problem by always using the implied value of the
certname converted to lower case.  We log an info message for the user
when we're not using the explicit value they've specified.  Info should
only display when using verbose mode to prevent log spam.