(maint) Fix installer to use new module_requirements
Before this patch the installer action expects the modules in
module_requirements to be two-item Arrays. This is no longer the case
and causes the installer action to fail.

This patch fixes this issue by updating the installer action to expect
the modules in module_requirements to be Hashes.

This patch does not include any changes in behavior.