(#11803) The uninstall action should use proper exit codes
Without this patch the uninstall action does not use proper exit codes.
The exit code of 0 is returned even when there are errors during the
uninstall process.

This patch fixes this issue by calling the `Kernel::exit` method with
the proper exit code after logging any errors during the uninstall
process. This change also comes with the added benefit of removing an
extra newline on stdout when errors are encountered, we now exit before
the Face when_rendering blocks gets to run.

This patch includes updated specs and acceptance tests related to this
change in behavior.