(#12126) When reloading files detect changes anywhere in the load path
Prior to this change, if a file is loaded, and a new version is placed at a
higher priority location in the load path, reload_changed would not detect it.
This makes it properly detected so the first copy in the load path at the time
reload_changed is called will always be loaded.