(#12631) Remove puppet module face
The current state of the module face was deemed not ready for shipping
because there was insufficient testing and formatting.  This removes the
parts of the code that make the module face accessible as a face
application, and the accompanying tests.  This change will be reverted
after the 2.7.11rc is cut so that work can continue on the module face
for the next release.