Fix range intersection monkey patch :& alias
I made a last minute change to the original pull request to not monkey
patch Range to have an intersection method if it was already defined.
However, I also added `unless method_defined? :intersection` to the &
alias after intersection had been defined, so the alias never got
defined.  This commit fixes the :& alias.