Fix a test that tries to create a dir with '*' in the name
Windows has a bigger list of reserved characters that can't be used in
directory names than Linux does, so a spec failed that tried to create
an invalid module dir actually created an invalid Windows dir.

http://msdn.microsoft.com/en-us/library/windows/desktop/aa365247(v=vs.85).aspx#naming_conventions

    The following reserved characters:
    < (less than)
    > (greater than)
    : (colon)
    " (double quote)
    / (forward slash)
    \ (backslash)
    | (vertical bar or pipe)
    ? (question mark)
    * (asterisk)

The main point of the failing spec was to show the restrictions on module
directory names, which is /^[-\w]+$/, so I'm just going to use an '=' in the
spec since Windows doesn't reserve that.