Make it so that modules with badly named dependencies don't blow up list
If a dependency name was an invalid module name, `puppet module list`
would blow up before it could list the unmet_dependencies because it
would attempt to instantiate a module with that name, and the
assert_validty method on the module class would raise.

Now we rescue that error so that we can still list unmet dependencies,
include the unmet dependency with the bad name.