Remove interrogate code
As a general command line principle we won't be prompting for user
input.  This means the interrogate code moved over with the module tool
can be deleted.

While deleting the interrogation code, other dead code was revealed.
There was an authenticate option to the http request in the forge code,
but it was never used.  It was referred to in the old module tool by
applications that were never finished and have since been removed
(registrar, releaser and unreleaser)