Put back the telly prototype logging
I must have messed up the rebase for this since this was in 2.7.x and
should have been in the geordi branch.  What I'm not not sure of is how
there were some many changes to the implementation, and almost none to
the specs but they all pass.  Must need more specs.