Fix spec ordering failure on environment
The module_requirements method was sorting the requirements, but only as
it went along, not for the entire list of requirements that got put in
the data structure.  This caused intermittent test failures since
sometimes the results would come out in a different order than the
specs asserted.

This moves the sort to happen *after* the data is all in.