(#11408) Skip default file permissions for sync'ed files on Windows
Previously the pluginsync acceptance test fails on Windows because we
were trying to guess the appropriate default user and group to create
sync'ed files as, and guessing wrong.

When using cygwin and ssh with publickey authentication, the cygwin
sshd server, running as `cyg_server` creates a token to impersonate
the ssh client, `Administrator` in my case, and starts a shell process
using the token. The new process runs as the authenticated user,
`Administrator`, but calls to `GetUserName` (which the sys-admin gem
calls), returns `cyg_server`. See "method 1: create a token from
scratch"[1].

As a result, puppet was setting the $vardir/lib directory owner to
`cyg_server`, group nobody, and mode 0755, which meant puppet running
as `Administrator` couldn't download (write) files into the directory.

The real issue, though is that we shouldn't be trying to guess what
the default owner/group for newly created files/directories that are
plugin sync'ed. This is because the default DACL varies across OS
(2003, 2008, ...) and user (Administrator vs SYSTEM).

This commit just omits the owner and group parameters from downloaded
file resources on Windows, so that they get created with the default
owner and group. The POSIX behavior is unchanged, using the
Process.uid/gid.

[1] http://cygwin.com/cygwin-ug-net/ntsec.html