Fix the modulepath dir interaction and add tests for it
The behaviors for interaction between modulepath and dir affect both
where a module gets installed, and what modules are considered for
constraint resolution.  This fixes some problematic interactions and add
tests to specify appropriate behavior.  The problem was that by
defaulting the install dir to the first path of the modulepath, it was
always present so it always got added to the front of the modulepath,
even when the modulepath was specified and the install dir wasn't.

There is also some changes to the way options are formatted for Puppet
help.  We're going with displaying options that take a value as

    optionname OPTIONVALUE

as opposed to

    optionname=

Either one is functionally equivalent, it will just affect the display
consistency.