(#7316) Initial acceptance tests
This is an acceptance test for #7316 and #6753.  It creates an application
and a face on the master, runs the agent, verifies that the pluginsync
brought down the correct files, and then attempts to run the app/face.

If the app/face run successfully then we've illustrated that the pluginsync
worked and that the app/face are available on the node, and also that the
face can be run without having a stub placeholder file in the application
directory (#6753).

These currently fail, which is why they are in the "pending" directory.