Unpacker always unpacks
I earlier removed the prompting and replaced it with a force option
check, but this made upgrades for dependencies not work.  If an install
doesn't need to happen, it shouldn't get to this point in the unpacker
anyway.

I'm pretty sure all this is reworked in Pieter's install branch anyway,
so this change may not matter.