Maint: Fix bad copy and paste
The code for creating the webrick access log file was copied from
ruby's `WEBrick::BasicLog.initialize` method so that we can prevent
the file descriptor from being inherited across an exec call. However,
the pasted version changed `sync = true` to just `sync`, which doesn't
do what we want.

This commit sets the sync mode on the access log file descriptor so
that writes are automatically flushed.