(#7801) Modify include function to accept arrays
When an array of classes is passed to the :include function, it
currently attempts to wrap the array in another array, and then pass the
entire array through as a single argument, causing an error.  This
commit calls flatten against the array before sending it through,
following the pattern of :realize, and enabling arrays to be passed as
well as comma-separated or individual class values.