Don't stub the database in Puppet::Rails tests.
The tests for the Puppet::Rails::Host and other Rails interface class stubbed
out parts of rails, in a way that confused things later on and ended up
breaking unrelated tests ... but only when they were invoked in a very
particular order.

Thankfully, just using a scratch database is both cheap and easy here, and
makes this all just work.  So, do that instead.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>