(#12844) Backwards compatibility for 'puppet agent --enable'
In version 2.7.10 we changed the name of the lock file that is
used by 'puppet agent --enable/--disable'.  A side effect is
that if the lockfile was created under an older version of
puppet, the "--enable" command in the newer versions won't
delete the old lockfile and thus you can't run your agents.
Make --enable attempt to detect and resolve this situation.