(#12254) Uninstall action uses console logger
Without this patch the uninstall action does not use the console logger
when printing error messages to the console. Error messages are being
sent directly to `$stderr`.

This patch resolves this issue by replacing calls appending error
messages to `$stderr` with calls to `Puppet.err`.

This patch does not introduce any changes in behavior and includes
updated accpetance tests.