(#12630) Run puppet agent as a service
Originally we were planning on installing puppet agent as a scheduled
task, but there are issues with this, such as the task scheduler service
itself may be disabled, may be configured to run as different user that
doesn't have sufficient privileges, etc.

This commit adds a wrapper script to run puppet agent as a service using
the win32-daemon gem. The Service Control Manager invokes the daemon.bat
script, which runs `rubyw.exe daemon.rb`.

The daemon ruby process sits in a loop executing `puppet agent --onetime` every
`runinterval` seconds, the value of which is obtained from puppet itself, and
can be configured in puppet.conf.

Windows services cannot write to stdout/err. Instead they are supposed to
write to the event log. But since we don't support that yet, the daemon
logs to a file in var/log/windows.log. By default the daemon logs at the
`notice` level.

In order to pass arguments to the service, you can either use the Service
Control Manager GUI or do the following from the command line:

  sc start puppet --debug --test --server foo.com

If --debug is passed as an argument, then the daemon will log at the
`debug` level.

This commit also updates the msi to install the daemon as a service. The
service's name is 'puppet' and its display name is 'Puppet Agent'. During
an install, msiexec will automatically stop the (old) service, install the
new files, and start the new service. The service is installed as an
automatic service, meaning it will start when Windows starts. It is
difficult to make this configurable from the command line, but can easily
be done using Orca/InstEd MSI editors.

Conflicts:

	wix/puppet.wxs