(maint) The --force install flag forces installation
Without this patch the `--force` flag for the install action does not
force the installation of modules that are already installed.

This patch resolves this issue by forcing the installation of modules
that are already installed. It should also be noted that using the
`--force` flag will implicitly ignore dependencies and only install the
module requested.