(#13027) Guard against invalid runinterval values
Previously, if the daemon was unable to start puppet, e.g. because the
wrong version of puppet was bundled into the installer, then the
runinterval value was getting set to 0, which caused the daemon to get
into an infinite loop launching puppet/ruby processes.

This commit guards against this, so that if runinterval is 0, we
instead default it to 1800 seconds, and display an error message in
the log.