Document the problem with the `apply_to_device` and related methods.
These methods were introduced, producing a pale shadow of the constraint
system that we already had, as part of building out device support.

Now that we are confident of the long term value of that support, and are
looking to better integrate it, we should start cleaning up these older relics
of the initial design.

This flags, in a comment, what the problem is and what the solution should be,
so that folks looking at the code get directed to the right sort of place.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>