Fix the shellquote parser function on Ruby 1.9
The semantics of `String#include?` changed a little in Ruby 1.9, so that you
can't just pass an arbitrary integer and have it check for character
containment.  (...or, perhaps, the semantics of characters vs integers
changed.  Either way, the effect is the same.)

This fixes the code to do the right thing on all platforms; we want to process
the string byte-wise, since that is how sh may well regard it, and we have to
work around some limitations on 1.8.5 semantics of string iterators.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>